@@ -6,7 +6,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.4</version>
+        <version>3.2.5</version>
     </parent>
 
     <groupId>fr.pinguet62</groupId>
@@ -15,8 +15,7 @@
     <packaging>pom</packaging>
 
     <properties>
-        <java.version>18</java.version>
-        <kafka.version>3.1.0</kafka.version><!-- fix Windows -->
+        <java.version>22</java.version>
     </properties>
 
     <modules>
@@ -34,8 +33,22 @@
         <module>spring-messagesource-database</module>
         <module>spring-mongo-embedded-args</module>
         <module>spring-security-jwt</module>
-        <module>spring-statemachine-sample</module>
         <module>spring-transactional</module>
         <module>wiremock-spring-boot</module>
     </modules>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>de.flapdoodle.embed</groupId>
+                <artifactId>de.flapdoodle.embed.mongo.spring30x</artifactId>
+                <version>4.6.2</version>
+            </dependency>
+            <dependency>
+                <groupId>com.github.npathai</groupId>
+                <artifactId>hamcrest-optional</artifactId>
+                <version>2.0.0</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
 </project>
