@@ -3,7 +3,7 @@ package fr.pinguet62.test.springdatathrowifnotfound.config;
 import org.springframework.data.jpa.repository.support.JpaEntityInformation;
 import org.springframework.data.jpa.repository.support.SimpleJpaRepository;
 
-import javax.persistence.EntityManager;
+import jakarta.persistence.EntityManager;
 
 public class ThrowIfNotFoundRepository<T, ID> extends SimpleJpaRepository<T, ID> implements ThrowIfNotFoundJpaRepository<T, ID> {
 
