@@ -3,7 +3,7 @@ package fr.pinguet62.test.springdatathrowifnotfound.config;
 import org.springframework.data.jpa.repository.support.JpaRepositoryFactory;
 import org.springframework.data.repository.core.RepositoryMetadata;
 
-import javax.persistence.EntityManager;
+import jakarta.persistence.EntityManager;
 
 public class ThrowIfNotFoundRepositoryFactory extends JpaRepositoryFactory {
 
