@@ -1,13 +1,13 @@
 package fr.pinguet62.test.jsr303;
 
+import jakarta.validation.ConstraintViolationException;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import org.external.ExternalModel;
 import org.hibernate.validator.constraints.Length;
 import org.junit.jupiter.api.Test;
 
-import javax.validation.ConstraintViolationException;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
 import java.util.ArrayList;
 import java.util.List;
 
