@@ -1,12 +1,11 @@
 package fr.pinguet62.mongo.embedded;
 
-import de.flapdoodle.embed.mongo.config.MongodConfig;
+import de.flapdoodle.embed.mongo.commands.MongodArguments;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 
 import java.util.Map;
 
-import static de.flapdoodle.embed.mongo.distribution.Version.Main.PRODUCTION;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.hasEntry;
@@ -15,8 +14,7 @@ class AppendArgsToMongodConfigBeanPostProcessorTest {
 
     @Test
     void shouldAppendArgsToExistingBean() {
-        MongodConfig existingMongodConfig = MongodConfig.builder()
-                .version(PRODUCTION)
+        MongodArguments existingMongodConfig = MongodArguments.builder()
                 .putArgs("initial", "preserved")
                 .putArgs("another", "also-preserved")
                 .build();
@@ -26,7 +24,7 @@ class AppendArgsToMongodConfigBeanPostProcessorTest {
                 Map.entry("second", "bar")));
         Object updatedMongodConfig = beanPostProcessor.postProcessAfterInitialization(existingMongodConfig, "mongodConfig");
 
-        assertThat(((MongodConfig) updatedMongodConfig).args(), allOf(
+        assertThat(((MongodArguments) updatedMongodConfig).args(), allOf(
                 hasEntry("initial", "preserved"),
                 hasEntry("another", "also-preserved"),
                 hasEntry("first", "foo"),
