@@ -9,7 +9,6 @@ import org.springframework.web.client.RestTemplate;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.springframework.http.HttpMethod.GET;
 import static org.springframework.test.context.TestExecutionListeners.MergeMode.MERGE_WITH_DEFAULTS;
 
 @WireMockApi(api = "facebook", propertyKey = "api.facebook.port") // mock configuration
@@ -32,9 +31,9 @@ class WireMockTest {
     final RestTemplate restTemplate = new RestTemplate();
 
     @Test
-    @WireMockCallMock(api = "facebook", method = GET, urlMatching = "/first", status = 200, body = "I'm first URL")
-    @WireMockCallMock(api = "facebook", method = GET, urlMatching = "/second", status = 200, bodyResource = "classpath:fr/pinguet62/springboot/wiremock/test.txt")
-    @WireMockCallMock(api = "google", method = GET, urlMatching = "/other", status = 200, body = "I'm the other")
+    @WireMockCallMock(api = "facebook", method = "GET", urlMatching = "/first", status = 200, body = "I'm first URL")
+    @WireMockCallMock(api = "facebook", method = "GET", urlMatching = "/second", status = 200, bodyResource = "classpath:fr/pinguet62/springboot/wiremock/test.txt")
+    @WireMockCallMock(api = "google", method = "GET", urlMatching = "/other", status = 200, body = "I'm the other")
     void test() {
         assertThat(restTemplate.getForEntity(facebookUrl + "/first", String.class).getBody(), is("I'm first URL"));
         assertThat(restTemplate.getForEntity(facebookUrl + "/second", String.class).getBody(), is("I'm second URL"));
